fix: spa crash
